Factor blob serialization into a helper
